chore: remove server_url support
It turns out the Last.fm library I'm using doesn't have support for changing the URL. I may make a PR or fork it so that it can have support for this. I plan to have a personal scrobble server instead and it would be nice for those who use Libre.fm or other scrobble servers (that being said, I have no clue where Libre.fm's scrobble feature is).